permisos

+++ resources/views/livewire/estadosLW.blade.php
@@ -6,9 +6,11 @@
     <div class="col-lg-12 mb-3">
         <div class="card">
             <div class="card-header" style="display:flex; align-items:right;">
-                <button type="button" wire:click='openModal("")' class="btn btn-outline-success">
-                    <i class="fa-solid fa-user-plus"></i>
-                </button>
+                @can("crear-estados")
+                    <button type="button" wire:click='openModal("")' class="btn btn-outline-success">
+                        <i class="fa-solid fa-user-plus"></i>
+                    </button>
+                @endcan
                 <select wire:model='cantidadEstados' class="form-select" style="width: 5%; margin-left: 5px; margin-right: 5px;">
                     <option value="5">5</option>
                     <option value="10">10</option>
@@ -34,8 +36,12 @@
                                 <td>{{$estado->getPais->nombre}}</td>
                                 <td class="text-center">
                                     <div class="btn-group" role="group" aria-label="Basic example">
-                                        <button wire:click='openModal({{$estado->id}})' type="button" class="btn btn-info"><i class="fa-solid fa-pen-to-square"></i></button>
-                                        <button wire:click='borrarEstadoModal({{$estado->id}})' type="button" class="btn btn-danger"><i class="fa-solid fa-trash"></i></button>
+                                        @can("editar-estados")
+                                            <button wire:click='openModal({{$estado->id}})' type="button" class="btn btn-info"><i class="fa-solid fa-pen-to-square"></i></button>
+                                        @endcan
+                                        @can("borrar-estados")
+                                            <button wire:click='borrarEstadoModal({{$estado->id}})' type="button" class="btn btn-danger"><i class="fa-solid fa-trash"></i></button>
+                                        @endcan
                                     </div>
                                 </td>
                             </tr>
